hit-type-agent/requirements.txt
agent-base==0.20.0dev0
agents-api==0.7.0.33
aio-pika==7.1.2
aiohttp==3.7.4.post0
aiormq==6.2.3
aiosignal==1.2.0
alert==0.0.2
anyio==3.5.0
asgiref==3.5.0
async-timeout==3.0.1
asynctest==0.13.0
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.4
click==8.1.2
data-source-api==0.23.0.12
distlib==0.3.4
filelock==3.6.0
flake8==4.0.1
frozenlist==1.3.0
googleapis-common-protos==1.53.0
grpcio==1.39.0
grpcio-health-checking==1.39.0
grpcio-reflection==1.39.0
grpcio-tools==1.39.0
h11==0.13.0
idna==3.2
iniconfig==1.1.1
isort==5.10.1
joblib==1.0.1
lz4==3.1.3
mccabe==0.6.1
multidict==5.1.0
numpy==1.21.1
packaging==21.3
pamqp==3.1.0
platformdirs==2.5.2
pluggy==1.0.0
protobuf==3.17.3
protocol-agents==0.18.0.90
psutil==5.9.0
py==1.11.0
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==3.0.8
pytest==7.1.1
pytest-asyncio==0.18.3
python-consul2==0.1.5
PyYAML==5.4.1
requests==2.26.0
sentry-sdk==1.5.4
six==1.16.0
sniffio==1.2.0
threadpoolctl==2.2.0
toml==0.10.2
tomli==2.0.1
tox==3.25.0
typing-extensions==3.10.0.0
Unidecode==1.2.0
urllib3==1.26.6
virtualenv==20.14.1
yarl==1.6.3
zipp==3.5.0
